refactor: DTO 패키지 이동 및 view에서 dto 객체 반환하도록 변경
