+++ src/main/java/chess/view/Command.java
@@ -1,8 +1,6 @@
 package chess.view;
 
-import chess.domain.board.Coordinate;
-import chess.domain.board.File;
-import chess.domain.board.Rank;
+import chess.dto.MovementRequest;
 import java.util.Arrays;
 
 public enum Command {
@@ -25,24 +23,11 @@ public enum Command {
                 .orElseThrow(() -> new IllegalArgumentException("잘못된 입력입니다."));
     }
 
-    public Coordinate sourceCoordinate(final InputTokens tokens) {
+    public MovementRequest movementRequest(final InputTokens tokens) {
         if (isNotMove()) {
             throw new UnsupportedOperationException("지원하지 않는 Command 기능입니다.");
         }
-        Coordinate source = mapToCoordinate(tokens.getSourceCoordinateToken());
-        return source;
-    }
-
-    public Coordinate targetCoordinate(final InputTokens tokens) {
-        if (isNotMove()) {
-            throw new UnsupportedOperationException("지원하지 않는 Command 기능입니다.");
-        }
-        Coordinate target = mapToCoordinate(tokens.getTargetCoordinateToken());
-        return target;
-    }
-
-    private Coordinate mapToCoordinate(final String input) {
-        return Coordinate.of(File.from(input.charAt(0)), Rank.from(input.charAt(1) - '0'));
+        return new MovementRequest(tokens.getSourceCoordinateToken(), tokens.getTargetCoordinateToken());
     }
 
     public boolean isStart() {
